+++ cmd/frontend/internal/auth/saml/user.go
@@ -12,59 +12,87 @@ import (
 	"github.com/sourcegraph/sourcegraph/pkg/actor"
 )
 
+type authnResponseInfo struct {
+	spec                 db.ExternalAccountSpec
+	email, displayName   string
+	unnormalizedUsername string
+	accountData          interface{}
+}
+
+func readAuthnResponse(p *provider, encodedResp string) (*authnResponseInfo, error) {
+	assertions, err := p.samlSP.RetrieveAssertionInfo(encodedResp)
+	if err != nil {
+		return nil, errors.WithMessage(err, "reading AuthnResponse assertions")
+	}
+	if wi := assertions.WarningInfo; wi.InvalidTime || wi.NotInAudience {
+		return nil, fmt.Errorf("invalid SAML AuthnResponse: %+v", wi)
+	}
+
+	pi, err := p.getCachedInfoAndError()
+	if err != nil {
+		return nil, err
+	}
+
+	firstNonempty := func(ss ...string) string {
+		for _, s := range ss {
+			if s := strings.TrimSpace(s); s != "" {
+				return s
+			}
+		}
+		return ""
+	}
+	attr := samlAssertionValues(assertions.Values)
+	email := firstNonempty(attr.Get("email"), attr.Get("emailaddress"))
+	if email == "" && mightBeEmail(assertions.NameID) {
+		email = assertions.NameID
+	}
+	if pn := attr.Get("eduPersonPrincipalName"); email == "" && mightBeEmail(pn) {
+		email = pn
+	}
+	info := authnResponseInfo{
+		spec: db.ExternalAccountSpec{
+			ServiceType: providerType,
+			ServiceID:   pi.ServiceID,
+			ClientID:    pi.ClientID,
+			AccountID:   assertions.NameID,
+		},
+		email:                email,
+		unnormalizedUsername: firstNonempty(attr.Get("login"), attr.Get("uid"), email),
+		displayName:          firstNonempty(attr.Get("displayName"), attr.Get("givenName")+" "+attr.Get("surname")),
+		accountData:          assertions,
+	}
+	if assertions.NameID == "" {
+		return nil, errors.New("the SAML response did not contain a valid NameID")
+	}
+	if info.email == "" {
+		return nil, errors.New("the SAML response did not contain an email attribute")
+	}
+	if info.unnormalizedUsername == "" {
+		return nil, errors.New("the SAML response did not contain a username attribute")
+	}
+	return &info, nil
+}
+
 // getOrCreateUser gets or creates a user account based on the SAML claims. It returns the
 // authenticated actor if successful; otherwise it returns an friendly error message (safeErrMsg)
 // that is safe to display to users, and a non-nil err with lower-level error details.
-func getOrCreateUser(ctx context.Context, p *provider, info *saml2.AssertionInfo) (_ *actor.Actor, safeErrMsg string, err error) {
+func getOrCreateUser(ctx context.Context, p *provider, info *authnResponseInfo) (_ *actor.Actor, safeErrMsg string, err error) {
 	var data db.ExternalAccountData
-	auth.SetExternalAccountData(&data.AccountData, info)
+	auth.SetExternalAccountData(&data.AccountData, info.accountData)
 
-	attr := samlAssertionValues(info.Values)
-	email := attr.Get("email")
-	if email == "" && mightBeEmail(info.NameID) {
-		email = info.NameID
-	}
-	login := attr.Get("login")
-	if login == "" {
-		login = attr.Get("uid")
-	}
-	displayName := attr.Get("displayName")
-	if displayName == "" {
-		displayName = attr.Get("givenName")
-	}
-	if displayName == "" {
-		displayName = login
-	}
-	if displayName == "" {
-		displayName = email
-	}
-	if displayName == "" {
-		displayName = info.NameID
-	}
-	if login == "" {
-		login = email
-	}
-	if login == "" {
-		return nil, "The SAML authentication provider did not contain an email attribute.", errors.New("SAML response did not contain email")
-	}
-	login, err = auth.NormalizeUsername(login)
+	username, err := auth.NormalizeUsername(info.unnormalizedUsername)
 	if err != nil {
-		return nil, fmt.Sprintf("Error normalizing the username %q. See https://about.sourcegraph.com/docs/config/authentication#username-normalization.", login), err
+		return nil, fmt.Sprintf("Error normalizing the username %q. See https://about.sourcegraph.com/docs/config/authentication#username-normalization.", info.unnormalizedUsername), err
 	}
 
 	userID, safeErrMsg, err := auth.CreateOrUpdateUser(ctx, db.NewUser{
-		Username:        login,
-		Email:           email,
-		EmailIsVerified: email != "", // TODO(sqs): https://github.com/sourcegraph/sourcegraph/issues/10118
-		DisplayName:     displayName,
+		Username:        username,
+		Email:           info.email,
+		EmailIsVerified: info.email != "", // TODO(sqs): https://github.com/sourcegraph/sourcegraph/issues/10118
+		DisplayName:     info.displayName,
 		// SAML has no standard way of providing an avatar URL.
 	},
-		db.ExternalAccountSpec{
-			ServiceType: providerType,
-			ServiceID:   p.ID().ID,
-			ClientID:    "",
-			AccountID:   info.NameID,
-		},
+		info.spec,
 		data,
 	)
 	if err != nil {